Design changes to client.get_player
Added an optional kwarg `node_id`.
Breaking change: positional arg `cls` to a kwarg.

`node_id` allows users to specify which node a new player should be created under. If the node cannot be found an error is raised.
Players that already exist are not affected by this kwarg.

`cls` is now a kwarg. This is a breaking change.