Changes to `Equalizer` which allow for construction without the `build()` classmethod.
Added `name` parameters to the `build()` classmethod and Equalizer constructor to allow setting a custom name.

Added a `name` property to Equalizer. Setting the name should be done by constrcutor or the `build()` classmethod

Added `__str__` and `__repr__` to Equalizer.

Fixed a bug where `set_eq` was setting a blank Equalizer class on the player.