Update player.py to work with dpy2.0
This adds a `**kwargs` consumption parameter to `Player.connect` and the subsequent call to method internals.

I opted for `**kwargs` both for backwards compatibility and inter-library usage.

The "fix" currently is to provide both `self_mute` and `self_deaf` kwargs to the method call. Please let me know if you'd like a better method.